android: Switch to our libhevc Switch to our fork of libhevc so we can patch CVE-2017-0472.  Google left this out of android-6.0.1_r79 :-(
android: Switch to our libmpeg2 and libnl
Switch to our fork of libmpeg2 so we can patch CVE-2017-0556
and CVE-2017-0557, and our fork of libnl so we can patch
CVE-2017-0553.

android: Switch to our freetype
Switch to our fork of freetype so we can patch CVE-2016-10244

asb-2017.05: fork platform/external/giflib

CVE-2016-5129: Fork of platform/external/v8

manifest: use Lineage fork of external/tremolo
The July ASB contains a fix in external/tremolo. We already have
a fork of this repo, but we currently track AOSP. There is one commit
missing in comparison to 6.0.1_r79, which will be pushed. Afterwards,
the July ASB change can be merged.

6.0.1_r79